pyproject.toml
[project]
name = "trp1-automaton-auditor"
version = "0.1.0"
description = "Automaton Auditor — Digital Courtroom agent graph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-groq>=0.2",
    "langchain-google-genai>=2.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pypdf>=4.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
